// src/types/entities.ts
export type AddressSnapshot = {
  formattedAddress: string;
  city: string;
  county: string;
  country: string;
  postalCode?: string | null;
  location: GeoPoint;
};